apps/api/pyproject.toml
[project]
name = "twype-api"
version = "0.1.0"
description = "Twype REST API"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "bcrypt>=4.3",
    "pgvector>=0.4.2",
    "python-jose[cryptography]>=3.4",
    "python-multipart>=0.0.20",
    "resend>=2.7",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn>=0.41.0",
    "pydantic[email]>=2.12.5",
]
